Encode HTML response as UTF-8
This is only tested with Python-3.  Some BibTeX responses have UTF-8
characters that are treated by Python-2 as ASCII and thus fail to
encode.  Python-3 treats the raw response as bytes, which we need to
encode in order to print.  This fix works for me on such entries, but
may not be the preferred strategy if including Python-2 support.